Renamed to DiscardedValues
